requirements.txt
aiogram==3.22.0
python-dotenv==1.0.0
aioodbc==0.5.0